+++ hw12_13_14_15_calendar/go.mod
@@ -3,9 +3,9 @@ module github.com/avtalabirchuk/otus-golang/hw12_13_14_15_calendar
 go 1.16
 
 require (
-	github.com/gorilla/mux v1.8.0 // indirect
-	github.com/sirupsen/logrus v1.8.1 // indirect
-	github.com/spf13/viper v1.9.0 // indirect
-	github.com/stretchr/testify v1.7.0 // indirect
-	golang.org/x/example v0.0.0-20210811190340-787a929d5a0d // indirect
+	github.com/gorilla/mux v1.8.0
+	github.com/jackc/pgx/v4 v4.13.0
+	github.com/sirupsen/logrus v1.8.1
+	github.com/spf13/viper v1.9.0
+	github.com/stretchr/testify v1.7.0
 )
