+++ hw12_13_14_15_calendar/internal/app/app_list_events_test.go
@@ -5,8 +5,9 @@ import (
 	"testing"
 	"time"
 
-	"github.com/avtalabirchuk/otus-golang/hw12_13_14_15_calendar/internal/storage"
 	"github.com/stretchr/testify/suite"
+
+	"github.com/avtalabirchuk/otus-golang/hw12_13_14_15_calendar/internal/storage"
 )
 
 type ListEventTest struct {
