+++ hw12_13_14_15_calendar/internal/server/http/responseWriter.go
@@ -1,0 +1,13 @@
+package internalhttp
+
+import "net/http"
+
+type responseWriter struct {
+	http.ResponseWriter
+	code int
+}
+
+func (w *responseWriter) WriteHeader(statusCode int) {
+	w.code = statusCode
+	w.ResponseWriter.WriteHeader(statusCode)
+}
