+++ hw12_13_14_15_calendar/internal/storage/memory/public.go
@@ -1,8 +1,6 @@
 package memorystorage
 
-import (
-	"github.com/avtalabirchuk/otus-golang/hw12_13_14_15_calendar/internal/storage"
-)
+import "github.com/avtalabirchuk/otus-golang/hw12_13_14_15_calendar/internal/storage"
 
 func New() storage.Storage {
 	result := store{}
